backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
tiktoken==0.5.2
numpy==1.24.3
scikit-learn==1.3.2
kafka-python==2.0.2
kubernetes==28.1.0
prometheus-client==0.19.0
structlog==23.2.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
psutil==5.9.6
clickhouse-driver==0.2.6 